blog-api-service/requirements.txt
fastapi == 0.104.1
uvicorn == 0.12.0
sqlmodel == 0.0.12
psycopg == 3.1.13
psycopg-binary == 3.1.13
python-keycloak == 3.7.0